requirements.txt
pyserial>=3.4
python-osc==1.7.7
requests